Refactored ChangelogCommand logic into ChangelogManager
